backend: move `reg` helper to assembly_printer.py
